enter

+++ controllers/videoController.js
@@ -43,8 +43,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -57,25 +60,37 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
+    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
   }
 };
 export const getEditVideo = async (req, res) => {
-  // 템플릿을 랜더링
+  const {
+    params: { id }
+  } = req;
+  const video = await Video.findById(id);
+  res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+};
+  
+  /* 템플릿을 랜더링
   const {
     params: { id }
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
 };
-
+*/
 export const postEditVideo = async (req, res) => {
   // video를 업데이트
   const {
@@ -94,10 +109,21 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
+  const video = await Video.findById(id);
+  await Video.findByIdAndRemove({ _id: id});
+  res.redirect(routes.home);
+};
+/*
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
 };
+*/
